add unique number generating for package guid
