+++ util/guid/guid_z_bench_test.go
@@ -14,19 +14,33 @@ import (
 )
 
 func Benchmark_S(b *testing.B) {
-	for i := 0; i < b.N; i++ {
-		guid.S()
-	}
+	b.RunParallel(func(pb *testing.PB) {
+		for pb.Next() {
+			guid.S()
+		}
+	})
 }
 
 func Benchmark_S_Data_1(b *testing.B) {
-	for i := 0; i < b.N; i++ {
-		guid.S([]byte("123"))
-	}
+	b.RunParallel(func(pb *testing.PB) {
+		for pb.Next() {
+			guid.S([]byte("123"))
+		}
+	})
 }
 
 func Benchmark_S_Data_2(b *testing.B) {
-	for i := 0; i < b.N; i++ {
-		guid.S([]byte("123"), []byte("456"))
-	}
+	b.RunParallel(func(pb *testing.PB) {
+		for pb.Next() {
+			guid.S([]byte("123"), []byte("456"))
+		}
+	})
+}
+
+func Benchmark_I(b *testing.B) {
+	b.RunParallel(func(pb *testing.PB) {
+		for pb.Next() {
+			guid.I()
+		}
+	})
 }
