+++ util/guid/guid_z_unit_test.go
@@ -11,6 +11,7 @@ package guid_test
 import (
 	"github.com/gogf/gf/container/gset"
 	"github.com/gogf/gf/util/guid"
+	"sync"
 	"testing"
 
 	"github.com/gogf/gf/test/gtest"
@@ -30,5 +31,27 @@ func Test_S(t *testing.T) {
 func Test_S_Data(t *testing.T) {
 	gtest.C(t, func(t *gtest.T) {
 		t.Assert(len(guid.S([]byte("123"))), 32)
+	})
+}
+
+func Test_I(t *testing.T) {
+	gtest.C(t, func(t *gtest.T) {
+		var (
+			set = gset.NewSet(true)
+			wg  = sync.WaitGroup{}
+			ch  = make(chan struct{})
+		)
+		wg.Add(10)
+		for i := 0; i < 10; i++ {
+			go func() {
+				<-ch
+				for i := 0; i < 100000; i++ {
+					t.Assert(set.AddIfNotExist(guid.I()), true)
+				}
+				wg.Done()
+			}()
+		}
+		close(ch)
+		wg.Wait()
 	})
 }
